+++ pkg/model/model.go
@@ -158,11 +158,17 @@ type MaskType struct {
 }
 
 type Masking struct {
-	Selector SelectorType `yaml:"selector"`
-	Mask     MaskType     `yaml:"mask,omitempty" jsonschema:"oneof_required=Mask"`
-	Masks    []MaskType   `yaml:"masks,omitempty" jsonschema:"oneof_required=Masks"`
-	Cache    string       `yaml:"cache,omitempty"`
-	Preserve string       `yaml:"preserve,omitempty"`
+	// Masking requires at least one Selector and one Mask definition.
+	// Case1: One selector, One mask
+	// Case2: One selector, Multiple masks
+	// Case3: Multiple selectors, One mask
+	// Case4: Multiple selectors, Multiple masks
+	Selector  SelectorType   `yaml:"selector,omitempty" jsonschema:"oneof_required=case1,oneof_required=case2"`
+	Selectors []SelectorType `yaml:"selectors,omitempty" jsonschema:"oneof_required=case3,oneof_required=case4"`
+	Mask      MaskType       `yaml:"mask,omitempty" jsonschema:"oneof_required=case1,oneof_required=case3"`
+	Masks     []MaskType     `yaml:"masks,omitempty" jsonschema:"oneof_required=case2,oneof_required=case4"`
+	Cache     string         `yaml:"cache,omitempty"`
+	Preserve  string         `yaml:"preserve,omitempty"`
 }
 
 type CacheDefinition struct {